+++ components/mp3/src/mp3_player.c
@@ -20,7 +20,7 @@ esp_err_t mp3_player_play_prev(mp3_player_handle_t player) {
     return player->play_prev(player);
 }
 
-esp_err_t mp3_player_play(mp3_player_handle_t player, int16_t idx) {
+esp_err_t mp3_player_play(mp3_player_handle_t player, uint16_t idx) {
     ESP_RETURN_ON_FALSE(player, ESP_ERR_INVALID_ARG, TAG, "invalid argument");
 
     return player->play(player, idx);
@@ -38,20 +38,20 @@ esp_err_t mp3_player_volume_decrease(mp3_player_handle_t player) {
     return player->volume_decrease(player);
 }
 
-esp_err_t mp3_player_volume(mp3_player_handle_t player, int16_t idx) {
+esp_err_t mp3_player_volume(mp3_player_handle_t player, uint8_t idx) {
     ESP_RETURN_ON_FALSE(player, ESP_ERR_INVALID_ARG, TAG, "invalid argument");
 
     return player->volume(player, idx);
 }
 
-esp_err_t mp3_player_eq(mp3_player_handle_t player, int16_t idx)
+esp_err_t mp3_player_eq(mp3_player_handle_t player, uint8_t idx)
 {
     ESP_RETURN_ON_FALSE(player, ESP_ERR_INVALID_ARG, TAG, "invalid argument");
 
     return player->eq(player, idx);
 }
 
-esp_err_t mp3_player_play_mode(mp3_player_handle_t player, int16_t idx)
+esp_err_t mp3_player_play_mode(mp3_player_handle_t player, uint8_t idx)
 {
     ESP_RETURN_ON_FALSE(player, ESP_ERR_INVALID_ARG, TAG, "invalid argument");
