+++ components/mp3/src/mp3_player.h
@@ -14,17 +14,17 @@ struct mp3_player_t {
 
     esp_err_t (*play_prev)(struct mp3_player_t *self);
 
-    esp_err_t (*play)(struct mp3_player_t *self, int16_t idx);
+    esp_err_t (*play)(struct mp3_player_t *self, uint16_t idx);
 
     esp_err_t (*volume_increase)(struct mp3_player_t *self);
 
     esp_err_t (*volume_decrease)(struct mp3_player_t *self);
 
-    esp_err_t (*volume)(struct mp3_player_t *self, int16_t idx);
+    esp_err_t (*volume)(struct mp3_player_t *self, uint8_t idx);
 
-    esp_err_t (*eq)(struct mp3_player_t *self, int16_t idx);
+    esp_err_t (*eq)(struct mp3_player_t *self, uint8_t idx);
 
-    esp_err_t (*play_mode)(struct mp3_player_t *self, int16_t idx);
+    esp_err_t (*play_mode)(struct mp3_player_t *self, uint8_t idx);
 
     esp_err_t (*stand_by)(struct mp3_player_t *self);
 
@@ -43,17 +43,17 @@ esp_err_t mp3_player_play_next(mp3_player_handle_t player);
 
 esp_err_t mp3_player_play_prev(mp3_player_handle_t player);
 
-esp_err_t mp3_player_play(mp3_player_handle_t player, int16_t idx);
+esp_err_t mp3_player_play(mp3_player_handle_t player, uint16_t idx);
 
 esp_err_t mp3_player_volume_increase(mp3_player_handle_t player);
 
 esp_err_t mp3_player_volume_decrease(mp3_player_handle_t player);
 
-esp_err_t mp3_player_volume(mp3_player_handle_t player, int16_t idx);
+esp_err_t mp3_player_volume(mp3_player_handle_t player, uint8_t idx);
 
-esp_err_t mp3_player_eq(mp3_player_handle_t player, int16_t idx);
+esp_err_t mp3_player_eq(mp3_player_handle_t player, uint8_t idx);
 
-esp_err_t mp3_player_play_mode(mp3_player_handle_t player, int16_t idx);
+esp_err_t mp3_player_play_mode(mp3_player_handle_t player, uint8_t idx);
 
 esp_err_t mp3_player_stand_by(mp3_player_handle_t player);
